tools/question-generator/requirements.txt
anthropic>=0.40.0
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
Jinja2>=3.1.0
python-dotenv>=1.0.0
